% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthcorpus.R
\docType{data}
\name{check_mapping}
\alias{check_mapping}
\title{Mappings for check data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 86 rows and 2 columns.
}
\usage{
check_mapping
}
\description{
Field mappings for the check data used for column names etc
\describe{
\item{\code{colname_en}}{character lowersnakecase column name}
\item{\code{desc_swe}}{character Swedish language explanation for field}
}
}
\keyword{datasets}
