% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_remap_colnames}
\alias{check_remap_colnames}
\title{Remaps column names using the internal check_mapping}
\usage{
check_remap_colnames(x)
}
\arguments{
\item{x}{character vector of column names}
}
\value{
character vector of new column names
}
\description{
Remaps column names using the internal check_mapping
}
