% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks_render_report}
\alias{checks_render_report}
\title{Render rmarkdown report with results from checks}
\usage{
checks_render_report(report = checks_report_path(), use_tmp = TRUE)
}
\arguments{
\item{report}{path to rmarkdown file, by default checks_report_path()}

\item{use_tmp}{boolean to indicate if resulting file should be placed in /tmp, default TRUE}
}
\description{
Render rmarkdown report with results from checks
}
