% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks_upload_report}
\alias{checks_upload_report}
\title{Upload file to S3}
\usage{
checks_upload_report(path, dest = "kthb/kthcorpus")
}
\arguments{
\item{path}{the full path to the file to upload}

\item{dest}{the path to the destination (s3 alias and bucket), by default "kthb/kthcorpus"}
}
\value{
exit code for upload command (0 means success)
}
\description{
Upload file to S3
}
\details{
Checks for minio client being present on the system and assumes
an alias "kthb" has been set up, as well as the bucket "kthcorpus"
which can be achieved by setting a MC_HOST_kthb environment variable
}
