% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\docType{data}
\name{countries_iso3}
\alias{countries_iso3}
\title{ISO3 codes and English country names from Wikipedia}
\format{
A data frame with 249 rows and 2 variables:
\describe{
\item{\code{country}}{character iso3 character code}
\item{\code{country_name}}{character country name description}
}
}
\usage{
countries_iso3
}
\description{
ISO3 codes and English country names from Wikipedia
}
\keyword{datasets}
