% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_funders}
\alias{cr_funders}
\title{Retrieve Funders from Crossref Funder Registry}
\usage{
cr_funders(location = "Sweden")
}
\arguments{
\item{location}{the geographical location, Default: 'Sweden'}
}
\value{
tibble with results
}
\description{
This function crawls the Crossref Funder Registry for entries in a specific location.
}
\examples{
\dontrun{
if(interactive()){
 cr_funders()
 }
}
}
