% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_funders_resolve}
\alias{cr_funders_resolve}
\title{Resolve funding organizations from Crossref Funding Registry from a text string}
\usage{
cr_funders_resolve(x)
}
\arguments{
\item{x}{a string vector of strings to lookup}
}
\value{
tibble with results
}
\description{
The CrossRef Open Funder Registry contains more than 30 000 funder names with DOI URIs
}
\examples{
\dontrun{
if(interactive()){
  cr_funders_resolve(c("rymdstyrelsen", "volvo", "saab"))
 }
}
}
