% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_publisher}
\alias{cr_publisher}
\title{Crossref lookup table for DOI prefixes}
\usage{
cr_publisher(doi_prefix = "all")
}
\arguments{
\item{prefix}{the prefix to look up, by default "all"}
}
\value{
data frame
}
\description{
Requests DOI prefixes from Crossref web site and returns as a data frame.
}
\examples{
\dontrun{
if(interactive()){
 cr_publisher()  # lookup all publishers
 cr_publisher(c("10.4172", "10.4236"))  # lookup given a few doi prefixes
 }
}
}
