% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_work_assertions}
\alias{cr_work_assertions}
\title{Crossref assertions associated with a DOI}
\usage{
cr_work_assertions(doi)
}
\arguments{
\item{doi}{character the DOI to look up information for}
}
\value{
data frame with assertions data for the DOI
}
\description{
For some DOIs, information like the conference_url can be
retrieved.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 # some DOIs do not have assertions...
 # "10.1145/263699"  |>
 "10.1007/978-3-031-26387-3_24" |>
  cr_work_assertions() |>
  select(c("doi", starts_with("conference")) |>
  glimpse()
 }
}
}
