% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_work_confinfo}
\alias{cr_work_confinfo}
\title{Crossref information about conferences, given a DOI}
\usage{
cr_work_confinfo(doi)
}
\arguments{
\item{doi}{character the DOI to look up information for}
}
\value{
data frame with conference information
}
\description{
Crossref Works API provides information about conference proceedings
for some DOIs.
}
\details{
For more information, see \url{# https://www.crossref.org/documentation/schema-library/markup-guide-record-types/conference-proceedings/}
}
\examples{
\dontrun{
if(interactive()){
 cr_work_confinfo("10.1145/263699") |> glimpse()
 }
}
}
