% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-diva.R
\docType{package}
\name{diva}
\alias{diva}
\alias{package-diva}
\title{KTH APIs}
\description{
DiVA is a Swedish research publication database with national and
institutional portals available. This R package downloads and makes
data available to use from R.
}
\details{
There is a national DiVA database portal and institutional database portals:
\itemize{
\item \href{https://www.diva-portal.org/smash/searchad.jsf}{National DiVA portal}
\item \href{https://kth.diva-portal.org/smash/search.jsf}{Institutional DiVA portal for KTH Royal Institute of Technology}
}
}
\keyword{package}
