% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_authors}
\alias{diva_authors}
\title{Retrieve DiVA authors for the default org from the DiVA portal}
\usage{
diva_authors()
}
\value{
data frame with results
}
\description{
This function returns parsed author information from DiVA data
}
