% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{diva_checks}
\alias{diva_checks}
\title{Data quality checks for DiVA publication data}
\usage{
diva_checks(authors, pubs, config = diva_config())
}
\arguments{
\item{authors}{a data frame with authors (from diva_download)}

\item{pubs}{a data frame with pubs (from diva_download)}

\item{config}{DiVA settings configuration, by default diva_config}
}
\value{
a list with slots for data frames with check results
}
\description{
This function makes a number of checks based on publications and author data
}
