% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_config}
\alias{diva_config}
\title{Configuration to use for DiVA client calls}
\usage{
diva_config()
}
\description{
By default the configuration uses settings for KTH.
}
\details{
Settings include the full base url to the DiVA portal for the institution,
the organisation id in DiVA (for example "177"), the organisation abbreviation
(for example "kth") and ybeg (2013), yend (2022) - the range of data to include.
}
