% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_meta}
\alias{diva_meta}
\title{Metadata for cached data files}
\usage{
diva_meta()
}
\value{
data frame with metadata including timestamp and age (in hours)
}
\description{
Metadata for cached data files
}
