% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_orcid_kthid_upload}
\alias{diva_orcid_kthid_upload}
\title{Upload DiVA ORCiD kthId pairs to object storage}
\usage{
diva_orcid_kthid_upload()
}
\description{
This function can be used to automatically download data from DiVA
from scratch, render the report and upload it to an S3 bucket
}
