% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_cora.R
\name{diva_organisations_cora}
\alias{diva_organisations_cora}
\title{Search DiVA organisations using CORA API}
\usage{
diva_organisations_cora(domain = "kth", freetext)
}
\arguments{
\item{domain}{string with domain, by default "kth"}

\item{freetext}{string with freetext to search for}
}
\value{
data frame with results
}
\description{
Search DiVA organisations using CORA API
}
