% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_pubs}
\alias{diva_pubs}
\title{Retrieve DiVA publications for the default org from the DiVA portal}
\usage{
diva_pubs(config = diva_config())
}
\arguments{
\item{config}{settings for DiVA client, by default using diva_config()}
}
\value{
data frame with results
}
\description{
This function sends a request to te DiVA portal from a CSV data export
covering publications in the default time period.
}
