% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_refresh}
\alias{diva_refresh}
\title{Refresh locally cached data files (and backup older data)}
\usage{
diva_refresh()
}
\value{
data frame with metadata including timestamp and age (in hours)
}
\description{
Refresh locally cached data files (and backup older data)
}
