% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_refresh_trigger}
\alias{diva_refresh_trigger}
\title{Refresh data trigger function}
\usage{
diva_refresh_trigger()
}
\description{
This function can be used to automatically download data from DiVA
from scratch, render the report and upload it to an S3 bucket
}
