% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_mods.R
\name{download_diva_mods_collection}
\alias{download_diva_mods_collection}
\title{Download a MODS collection XML file from DiVA given parameters}
\usage{
download_diva_mods_collection(
  y_beg = 2023,
  y_end = 2023,
  orgid = 177,
  destfile = file.path(tempdir(), "diva-mods-collection.xml"),
  publication_types = c("bookReview", "review", "article", "conferencePaper")
)
}
\arguments{
\item{y_beg}{from year issued (inclusive)}

\item{y_end}{to year issued (inclusive)}

\item{orgid}{the organisation id in DiVA, by default 177 (for KTH)}

\item{destfile}{the path to the xml file, by default a file in the tempdir}

\item{publication_types}{a character vector of publication types}
}
\description{
Download a MODS collection XML file from DiVA given parameters
}
