% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vies.R
\name{funders_eu_search}
\alias{funders_eu_search}
\title{Search Funding & tender opportunities}
\usage{
funders_eu_search(
  legal_name = "Kungliga Tekniska",
  country = NULL,
  reg_number = NULL,
  pic = NULL,
  vat = NULL,
  erasmus_code = NULL
)
}
\arguments{
\item{legal_name}{the legal name search string}

\item{country}{the country search string}

\item{reg_number}{the registration number search string}

\item{pic}{the Participant Identification Code (9 digits)}

\item{vat}{the VAT number}

\item{erasmus_code}{the Erasmus code}
}
\value{
tibble with results
}
\description{
The EU Single Electronic Data Interchange Area (SEDIA) offers a search API
}
\details{
\url{https://ec.europa.eu/info/funding-tenders/opportunities/portal/screen/how-to-participate/participant-register-search}
}
\examples{
\dontrun{
if(interactive()){
 funders_eu_search(legal_name = "Kungliga Tekniska")
 funders_eu_search(vat = "SE202100305401")
 }
}
}
