% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrplus.R
\name{hr_latest}
\alias{hr_latest}
\title{List contents of minio S3 bucket}
\usage{
hr_latest(bucket = "hrplus")
}
\arguments{
\item{bucket}{name of bucket to list}
}
\description{
List contents of minio S3 bucket
}
