% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrplus.R
\name{hr_ls}
\alias{hr_ls}
\title{List contents of minio S3 bucket}
\usage{
hr_ls(bucket = "hrplus")
}
\arguments{
\item{bucket}{name of bucket to list}
}
\description{
Before using this function, set up credentials for accessing minio using S3
}
\details{
For example use environment variables like these:
#AWS_ACCESS_KEY_ID=supersecret
#AWS_SECRET_ACCESS_KEY=supersecret
#AWS_S3_ENDPOINT=lib.kth.se:9000
#AWS_DEFAULT_REGION=bibliometrics
This can be achieved by adding these lines to the ~/.Renviron file.
#file.edit("~/.Renviron")
#readRenviron("~/.Renviron")
}
