% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-diva.R
\docType{data}
\name{hr_mapping}
\alias{hr_mapping}
\title{HR mapping}
\format{
A data frame with 14 rows and 3 variables:
\describe{
\item{\code{name}}{character Name of column}
\item{\code{table}}{character Source table}
\item{\code{field}}{character Source field}
}
}
\usage{
hr_mapping
}
\description{
Field mappings used in HR data extract
}
\details{
DETAILS
}
\keyword{datasets}
