% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrplus.R
\name{hr_plus}
\alias{hr_plus}
\title{Parse content from HR-plus data export}
\usage{
hr_plus(bucket = "hrplus", file, offset)
}
\arguments{
\item{bucket}{name of bucket to list}

\item{file}{optional character string for file name in bucket}

\item{offset}{optional offset in days (when retrieving older files)}
}
\description{
Parse content from HR-plus data export
}
