% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrplus.R
\name{hr_read_csv}
\alias{hr_read_csv}
\title{Parse and read the HR data in CSV format}
\usage{
hr_read_csv(file)
}
\arguments{
\item{file}{path to file}
}
\description{
Parse and read the HR data in CSV format
}
