% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthcorpus-package.R
\docType{data}
\name{kth_besta}
\alias{kth_besta}
\title{Employment Titles from KTH according to BESTA}
\format{
A data frame with 53 rows and 3 variables:
\describe{
\item{\code{emp_code}}{character code for employment title}
\item{\code{emp_title_sv}}{character title in mixed case Swedish}
\item{\code{emp_title_en}}{character title in mixed case English}
}
}
\usage{
kth_besta
}
\description{
Employment Titles have BESTA codes and descriptions in English and Swedish
}
\keyword{datasets}
