% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{kth_diva_authors}
\alias{kth_diva_authors}
\title{Retrieve DiVA authors for KTH from the KTH DiVA portal}
\usage{
kth_diva_authors(use_cache = TRUE, refresh_cache = FALSE)
}
\arguments{
\item{use_cache}{logical flag to indicate locally cached data should be used,
default: TRUE}

\item{refresh_cache}{logical flag to indicate if local data cache should be
refreshed, default: FALSE}
}
\value{
data frame with results
}
\description{
This function returns parsed author information from DiVA data
}
