% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{kth_diva_checks}
\alias{kth_diva_checks}
\title{Data quality checks for DiVA publication data}
\usage{
kth_diva_checks()
}
\value{
a list with slots for data frames with check results
}
\description{
This function makes a number of checks based on publications and author data
}
