% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curated.R
\name{kth_diva_curated}
\alias{kth_diva_curated}
\title{Curated DiVA corpus data}
\usage{
kth_diva_curated(jq = NULL)
}
\arguments{
\item{jq}{query for \href{https://docs.ropensci.org/jqr/}{filtering the data}}
}
\value{
tibble
}
\description{
Curated DiVA corpus data
}
\examples{
\dontrun{
 kth_diva_curated()
}
}
