% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{kth_diva_pubs}
\alias{kth_diva_pubs}
\title{Retrieve DiVA publications for KTH from the KTH DiVA portal}
\usage{
kth_diva_pubs(use_cache = TRUE, refresh_cache = FALSE)
}
\arguments{
\item{use_cache}{logical flag to indicate locally cached data should be used,
default: TRUE}

\item{refresh_cache}{logical flag to indicate if local data cache should be
refreshed, default: FALSE}
}
\value{
data frame with results
}
\description{
This function sends a request to KTH's DiVA portal from a CSV data export
covering KTH publications between 2013 and 2022.
}
