% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{kth_diva_pubs_deprecated}
\alias{kth_diva_pubs_deprecated}
\title{Retrieve DiVA publications for KTH from the KTH DiVA portal}
\usage{
kth_diva_pubs_deprecated(
  orgid = "177",
  year_beg = "2013",
  year_end = "2022",
  use_cache = TRUE
)
}
\arguments{
\item{orgid}{the DiVA organisation id, by default "177" for KTH}

\item{year_beg}{the beginning of the period, by default "2013"}

\item{year_end}{the end of the period, by default "2022"}

\item{use_cache}{logical flag to indicate if cached data should be used, default: TRUE}
}
\value{
data frame with results
}
\description{
This function sends a request to KTH's DiVA portal for a CSV data export
covering KTH publications between 2013 and 2022.
}
