% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthcorpus.R
\docType{data}
\name{kth_employment_title}
\alias{kth_employment_title}
\title{Employment Titles from KTH}
\format{
A data frame with 61 rows and 5 variables:
\describe{
\item{\code{emp_code_hr}}{character code for employment title}
\item{\code{emp_desc}}{character description in uppercase Swedish}
\item{\code{title_sv}}{character title in mixed case Swedish}
\item{\code{title_en}}{character title in mixed case English}
\item{\code{title_comment}}{character comment about this category of employees}
}
}
\usage{
kth_employment_title
}
\description{
Employment Titles have codes and descriptions in English and Swedish
}
\keyword{datasets}
