% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{kth_issues_pubauth}
\alias{kth_issues_pubauth}
\title{Issues with publication and author affiliations}
\usage{
kth_issues_pubauth(pass = Sys.getenv("DIVA_PASS"), jq = NULL)
}
\arguments{
\item{pass}{passphrase required for accessing the data,
Default: Sys.getenv("DIVA_PASS")}

\item{jq}{query for \href{https://docs.ropensci.org/jqr/}{filtering the data}}
}
\value{
json object
}
\description{
Issues with publication and author affiliations
}
\examples{
\dontrun{
 kth_issues_pubauth()
}
}
\seealso{
\code{\link[rappdirs]{app_dir}}
\code{\link[rcrypt]{decrypt}}
\code{\link[readr]{read_file}}
}
