% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_institutions.R
\name{latest_ror_url}
\alias{latest_ror_url}
\title{Get latest Zenodo ROR filename}
\usage{
latest_ror_url(
  url = "https://zenodo.org/api/records/?communities=ror-data&sort=mostrecent"
)
}
\arguments{
\item{url}{the URL for Zenodo ROR data}
}
\description{
Get latest Zenodo ROR filename
}
