% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{linkify}
\alias{linkify}
\title{Convert a string with a URL to a link}
\usage{
linkify(
  href,
  text = shorten(href),
  title = href,
  target = c("href", "PID", "titlesearch", "freetextsearch", "DOI", "ScopusID", "ISSN",
    "ISBN", "ISI", "ORCID")
)
}
\arguments{
\item{href}{the link}

\item{text}{the text to show for the link, by default shortened href}

\item{title}{the tooltip, by default the href}

\item{target}{the type of link to make}
}
\description{
Convert a string with a URL to a link
}
