% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minio_client.R
\name{minio_get}
\alias{minio_get}
\title{Get a file from an S3 bucket}
\usage{
minio_get(file, bucket = "bibliometrics")
}
\arguments{
\item{file}{the file}

\item{bucket}{the bucket to use, default is "bibliometrics"}
}
\description{
Get a file from an S3 bucket
}
