% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minio_client.R
\name{minio_ls}
\alias{minio_ls}
\title{List files in an S3 bucket}
\usage{
minio_ls(bucket = "bibliometrics")
}
\arguments{
\item{bucket}{the bucket to use, default is "bibliometrics"}
}
\description{
List files in an S3 bucket
}
