% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_mods.R
\name{mods_collection_file_to_tbl}
\alias{mods_collection_file_to_tbl}
\title{Convert a mods collection .xml file to a tibble and
extract some external identifiers etc}
\usage{
mods_collection_file_to_tbl(mods_file)
}
\arguments{
\item{mods_file}{file path to downloaded diva mods collection xml file}
}
\description{
Convert a mods collection .xml file to a tibble and
extract some external identifiers etc
}
