% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_extract.R
\name{neo4j_header_nodes}
\alias{neo4j_header_nodes}
\title{Assemble a header for loading data into neo4j}
\usage{
neo4j_header_nodes(
  data,
  id_field,
  array_fields,
  ignore_fields = NULL,
  id_namespace = NULL,
  label_field
)
}
\arguments{
\item{data}{a data frame to be exported}

\item{id_field}{the name of the primary key field}

\item{array_fields}{a vector of field names for fields that contain arrays}

\item{ignore_fields}{a vector of field names to exclude, Default: NULL}

\item{id_namespace}{an identifier namespace in neo4j parlance, Default: NULL}

\item{label_field}{the name of the field containing labels}
}
\value{
a string with header info that can be used by neo4j imports
}
\description{
Assemble a header for loading data into neo4j
}
\examples{
\dontrun{
if(interactive()){
 neo4j_header_nodes(kth_diva_pubs(), id_field = "PID", label_field = "Title")
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{slice}}
\code{\link[tidyselect]{all_of}}
}
