% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_extract.R
\name{neo4j_header_relations}
\alias{neo4j_header_relations}
\title{Assemble a header for loading edge data into neo4j}
\usage{
neo4j_header_relations(
  data,
  relation_type_field,
  relation_src_field,
  relation_dst_field,
  ignore_fields = NULL,
  array_fields
)
}
\arguments{
\item{data}{a data frame to be exported}

\item{relation_type_field}{field name with values describing the relationship type}

\item{relation_src_field}{field name with id values for the node where the relation starts}

\item{relation_dst_field}{field name with id values for the node where the relation ends}

\item{ignore_fields}{logical to indicate which fields should be ignored (if any), Default: NULL}

\item{array_fields}{a vector of field names for fields that contain arrays}
}
\value{
a string with header info that can be used by neo4j imports
}
\description{
Assemble a header for loading edge data into neo4j
}
\examples{
\dontrun{
if(interactive()){
 neo4j_header_edges(data)
 }
}
}
