% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_mods.R
\name{read_diva_csvall2}
\alias{read_diva_csvall2}
\title{Read rectangular data from DiVA given a specific identifier}
\usage{
read_diva_csvall2(pid)
}
\arguments{
\item{pid}{the DiVA publication identifier}
}
\description{
Read rectangular data from DiVA given a specific identifier
}
