% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_mods.R
\name{read_diva_mods}
\alias{read_diva_mods}
\title{Read a MODS file from DiVA given a DiVA publication identifier}
\usage{
read_diva_mods(pid)
}
\arguments{
\item{pid}{DiVA publication identifier}
}
\description{
Read a MODS file from DiVA given a DiVA publication identifier
}
