% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthcorpus.R
\docType{data}
\name{research_areas}
\alias{research_areas}
\title{Research Subject Areas from https://uka.se}
\format{
A data frame with 308 rows and 4 variables:
\describe{
\item{\code{id}}{integer subject area code}
\item{\code{level}}{double a level, granularity for code (1..3)}
\item{\code{swe}}{character description in Swedish}
\item{\code{eng}}{character description in English}
}
}
\usage{
research_areas
}
\description{
Research Subject Areas have codes, descriptions and are arranged
in three levels, 6 at level 1, 42 at level 2, 260 at level 3 (5 digit codes)
See \url{https://www.uka.se/statistik--analys/information-om-statistiken/amneslistor-och-huvudomraden/2017-02-14-forskningsamnen.html}
}
\details{
Data from \url{https://www.uka.se/download/18.7391c377159bc0155b81ef8/1487841861615/forskningsamnen-standard-2011.xlsx}
}
\keyword{datasets}
