% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_abstract_extended}
\alias{scopus_abstract_extended}
\title{Request extended information from Scopus Abstract API}
\usage{
scopus_abstract_extended(sid)
}
\arguments{
\item{sid}{ScopusID such as "SCOPUS_ID:85140569271" or "85140569271"}
}
\value{
list with three slots for abstract, authorgroups and correspondence
}
\description{
The response includes the abstract itself, associated author groups with
data including source text for affiliated organisations, and affiliated
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 mysid <- scopus$publications$`dc:identifier`[1]
 scopus_abstract_extended(mysid)
 }
}
}
\seealso{
\code{\link[httr]{content}}
\code{\link[readr]{parse_atomic}}
}
