% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_browse}
\alias{scopus_browse}
\title{Browse a Scopus electronic identifier}
\usage{
scopus_browse(eid)
}
\arguments{
\item{eid}{Electronic Identifier}
}
\description{
Browse a Scopus electronic identifier
}
