% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_classify}
\alias{scopus_classify}
\title{Classify subject categories given a Scopus identifier}
\usage{
scopus_classify(sid, scopus = scopus_from_minio())
}
\arguments{
\item{sid}{a scopus identifier}

\item{scopus}{by default publication given by scopus_from_minio()}
}
\value{
a tibble with results from SwePubs classification API
}
\description{
Classify subject categories given a Scopus identifier
}
