% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_config}
\alias{scopus_config}
\title{Configuration for credentials when using Scopus API}
\usage{
scopus_config(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to display messages for missing environment varibles,
by default FALSE}
}
\description{
Use .Renviron to provide credentials for API access.
}
\details{
Set values for SCOPUS_API_KEY and optionally for SCOPUS_API_TOKEN if making
requests from outside of institutional IPs linked to the subscriber.\preformatted{ file.edit("~/.Renviron")
 readRenviron("~/.Renviron")
}
}
