% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_filter_kth_authors}
\alias{scopus_filter_kth_authors}
\title{Authors affiliated with KTH for a default search request}
\usage{
scopus_filter_kth_authors(scopus = scopus_search_pubs_kth())
}
\arguments{
\item{scopus}{by default the result from scopus_search_pubs_kth()}
}
\description{
Authors affiliated with KTH for a default search request
}
