% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_mods.R
\name{scopus_mods}
\alias{scopus_mods}
\title{Generate MODS from a Scopus identifier}
\usage{
scopus_mods(sid, scopus = scopus_from_minio(), ko = kthid_orcid())
}
\arguments{
\item{sid}{Scopus identifier}

\item{scopus}{reference data for the identifier, by default from scopus_from_minion()}

\item{ko}{reference data for known KTH author identifiers' ORCiD associations
by default from kthid_orcid()}
}
\description{
Generate MODS from a Scopus identifier
}
