% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_mods.R
\name{scopus_mods_crawl}
\alias{scopus_mods_crawl}
\title{Generate MODS from several Scopus identifiers}
\usage{
scopus_mods_crawl(sids, scopus = scopus_from_minio(), ko = kthid_orcid())
}
\arguments{
\item{sids}{Scopus identifier}

\item{scopus}{reference data for the identifier, by default from scopus_from_minion()}

\item{ko}{reference data for known KTH author identifiers' ORCiD associations
by default from kthid_orcid()}
}
\description{
Generate MODS from several Scopus identifiers
}
