% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_mods.R
\name{scopus_mods_params}
\alias{scopus_mods_params}
\title{Creates MODS for a given Scopus identifier}
\usage{
scopus_mods_params(scopus, sid, kthid_orcid_lookup = kthid_orcid())
}
\arguments{
\item{scopus}{the result from scopus_search_pubs_kth()}

\item{sid}{a scopus identifier (required for getting extended abstract reference data)}

\item{kthid_orcid}{a lookuptable of orcids with a known kthid associated from kthid_orcid() fcn}
}
\value{
an object with parameters which can be used to generate MODS with the
create_diva_mods() function
}
\description{
Creates MODS for a given Scopus identifier
}
