% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_ratelimit_quota}
\alias{scopus_ratelimit_quota}
\title{Scopus API ratelimit quota information}
\usage{
scopus_ratelimit_quota()
}
\value{
a list with the information from the headers
}
\description{
Makes a request and displays the headers containing rate limit information
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{X-RateLimit-Limit       <----Shows API quota setting
X-RateLimit-Remaining   <----Shows API remaining quota
X-RateLimit-Reset       1234567891 <----Date/Time in Epoch seconds when API quota resets
}\if{html}{\out{</div>}}
}
