% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_req_abstract}
\alias{scopus_req_abstract}
\title{Fetch abstract given Scopus ID}
\usage{
scopus_req_abstract(
  sid,
  endpoint = "https://api.elsevier.com/content/abstract/scopus_id/"
)
}
\arguments{
\item{sid}{ScopusID}

\item{endpoint}{the endpoint to use for the request (default value provided)}
}
\description{
Fetch abstract given Scopus ID
}
\details{
see https://dev.elsevier.com/documentation/AbstractRetrievalAPI.wadl
}
\examples{
\dontrun{
  scopus_req_abstract("SCOPUS_ID:85140569271")
}
}
