% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_search_aut_kth}
\alias{scopus_search_aut_kth}
\title{Authors affiliated with KTH for a default search request}
\usage{
scopus_search_aut_kth()
}
\description{
Authors affiliated with KTH for a default search request
}
