% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_client.R
\name{scopus_search_id}
\alias{scopus_search_id}
\title{Retrieve publications from Scopus Search API for specified Scopus ids}
\usage{
scopus_search_id(eid)
}
\arguments{
\item{eid}{character vector of scopus eids, such as "2-s2.0-85144114333"}
}
\description{
Note: For long vectors of scopus eids, a warning will be issued
if the query generates a long query string parameter that may
run into issues with recommendations regarding RFC 2616.
}
