% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthcorpus-package.R
\docType{data}
\name{ss_employment_title}
\alias{ss_employment_title}
\title{Employment Titles from Statistics Sweden}
\format{
A data frame with 7626 rows and 4 variables:
\describe{
\item{\code{id}}{character code for employment title}
\item{\code{desc_swe}}{character description in Swedish}
\item{\code{cat_desc}}{character staff category}
\item{\code{is_uf_ta}}{character UF denotes educational or research title, TA denotes technical or administrative}
}
}
\usage{
ss_employment_title
}
\description{
Employment Titles have codes and descriptions
}
\details{
See \url{https://anstallningsbenamningar.scb.se/}
}
\keyword{datasets}
