% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_institutions.R
\name{tidy_rorzip}
\alias{tidy_rorzip}
\title{Get ROR data from Zenodo}
\usage{
tidy_rorzip(url = latest_ror_url())
}
\arguments{
\item{url}{the URL for ROR data from Zenodo}
}
\description{
Get ROR data from Zenodo
}
