% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthcorpus.R
\docType{data}
\name{unit_names}
\alias{unit_names}
\title{DiVA unit identifiers for KTH with Swedish and English translations}
\format{
A data frame with 160 rows and 6 variables:
\describe{
\item{\code{orgid}}{integer representing DiVA organizational unit id}
\item{\code{unit_abbr}}{character KTH organizational abbreviation code}
\item{\code{level}}{integer representing node level where 0 is the root node}
\item{\code{unit_en}}{character title in mixed case English}
\item{\code{unit_sv}}{character title in mixed case Swedish}
\item{\code{is_closed}}{logical denoting if mapping is no longer in use}
}
}
\usage{
unit_names
}
\description{
In DiVA, organizational units have identifiers. This mapping table
maps these units in DiVA to abbreviated unit codes used within KTH and provides
translations in English and Swedish, as well as organization level with higher
numbers representing a deeper nesting within the organizational structure (0 = root)
}
\keyword{datasets}
