% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_institutions.R
\name{update_ror_lookup_minio}
\alias{update_ror_lookup_minio}
\title{Save ROR lookup data to Minio}
\usage{
update_ror_lookup_minio(ror = tidy_rorzip())
}
\arguments{
\item{ror}{the ROR dataset as collected from Zenodo with \code{tidy_rorzip()}}
}
\description{
Save ROR lookup data to Minio
}
