% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vies.R
\name{vat_parse}
\alias{vat_parse}
\title{Parses a European VAT number}
\usage{
vat_parse(x)
}
\arguments{
\item{x}{the VAT number}
}
\value{
a list with slots for the segments
}
\description{
This function validates an European VAT number by checking
the length of the string and whether it begins with two
capital letters and is followed by 12 digits
}
\examples{
\dontrun{
if(interactive()){
 vat_parse("SE202100305401")
 }
}
}
