% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vies.R
\name{vat_resolve_vies}
\alias{vat_resolve_vies}
\title{Resolve vatNumber from European Union VIES REST API}
\usage{
vat_resolve_vies(x)
}
\arguments{
\item{x}{a string with a properly formatted European VAT number}
}
\value{
tibble with results (such as address, organization name etc)
}
\description{
The VIES REST API, see \url{https://viesapi.eu/test-vies-api/}, allows for online
validation of European VAT numbers
}
\examples{
\dontrun{
if(interactive()){
 vat_resolve_vies("SE202100305401")
 }
}
}
