% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_read.R
\name{wos_read_tsv}
\alias{wos_read_tsv}
\title{Read Web Of Science tab separated file(s)}
\usage{
wos_read_tsv(file, format = c("unnested", "compact"))
}
\arguments{
\item{file}{one or more paths to wos tsv files}

\item{format}{one of "unnested" (default) or "compact"}
}
\value{
either one tibble (compact) or a list of several unnested tibbles
}
\description{
Read one or more wos tsv files
}
