% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthcorpus.R
\docType{data}
\name{wos_tag_fields}
\alias{wos_tag_fields}
\title{Web of Science tags and field names}
\format{
A data frame with 101 rows and 4 variables:
\describe{
\item{\code{field}}{character suggested field name using snake case}
\item{\code{tag}}{character short tag or field abbreviation used in WoS}
\item{\code{tag_long}}{character a slightly longer explanation}
\item{\code{tag_set}}{character the kind of tag, a grouping}
}
}
\usage{
wos_tag_fields
}
\description{
A curated set of WoS tags and suggested field names, explanations.
}
\details{
DETAILS
}
\keyword{datasets}
