% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_mods.R
\name{write_mods_chunked}
\alias{write_mods_chunked}
\title{Write MODS collections as XML files, where each contains a certain number of MODS}
\usage{
write_mods_chunked(mods, outdir = ".", prefix = "mods", chunk_size = 25)
}
\arguments{
\item{mods}{a vector of MODS generated from scopus_mods_crawl()}

\item{outdir}{the directory to write the files}

\item{prefix}{the pattern to prefix file names with, by default "mods"}

\item{chunk_size}{the number of MODS in each batch, by default 25}
}
\description{
Write MODS collections as XML files, where each contains a certain number of MODS
}
