% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_client.R
\name{diva_download}
\alias{diva_download}
\title{Download publication or author data from DiVA}
\usage{
diva_download(
  set = c("pub", "aut"),
  org_id = "177",
  year_beg = 2013,
  year_end = 2022,
  sync = FALSE
)
}
\arguments{
\item{set}{dataset to use, either "pub" (default) or "aut"}

\item{org_id}{organisation to filter for, by default "177" (KTH)}

\item{year_beg}{first year in a range of years, default 2013}

\item{year_end}{last year in a range of years, default 2022}

\item{sync}{logical to indicate if sync should be made to s3 (default: FALSE)}
}
\description{
Download publication or author data from DiVA
}
