% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swepub_client.R
\name{swepub_checks}
\alias{swepub_checks}
\title{Checks from swepub processing of KTH publication records
This function calls an API at SwePub for KTH data in a specific time period}
\usage{
swepub_checks(
  config = diva_config(),
  year_beg = 1900L,
  year_end = lubridate::year(Sys.time())
)
}
\arguments{
\item{config}{the configuration to use, by default diva_config (w org "kth")}

\item{year_beg}{the beginning year, default 1900}

\item{year_end}{the ending year, default is the current year}
}
\value{
tibble with results
}
\description{
Checks from swepub processing of KTH publication records
This function calls an API at SwePub for KTH data in a specific time period
}
\examples{
\dontrun{
if(interactive()){
 swepub_checks()
 }
}
}
